[#21] Entrance logic added (step6: successful integration for user's defaults!!)
